+++ lib/repositories/user/local_user_source.dart
@@ -7,8 +7,12 @@ abstract class LocalUserSource {
   /// allowing it to be retrieved later.
   Future<void> saveWallet(String mnemonic);
 
-  /// Returns the [Wallet] instance associated with the current
-  /// application user.
+  /// Returns the mnemonic that is associated to the current application user.
+  /// If no mnemonic has been saved yet, returns `null`.
+  Future<List<String>> getMnemonic();
+
+  /// Returns the [Wallet] instance of the current application user.
+  /// If no [Wallet] instance has been saved yet, returns `null`.
   Future<Wallet> getWallet();
 
   /// Saves the given [data] as the current local user data.