+++ lib/ui/screens/export.dart
@@ -3,6 +3,8 @@ export 'home_screen.dart';
 export 'login_screen.dart';
 export 'post_details_screen.dart';
 export 'recover_account_screen.dart';
+export 'security_login_biometrics_screen.dart';
+export 'security_login_password_screen.dart';
 export 'security_set_biometrics_screen.dart';
 export 'security_set_password_screen.dart';
 export 'splash_screen.dart';