+++ lib/ui/widgets/mnemonic/export.dart
@@ -1,0 +1,1 @@
+export 'mnemonic_visualizer.dart';