+++ lib/usecases/account/account.dart
@@ -3,6 +3,7 @@ export 'usecase_check_login.dart';
 export 'usecase_generate_mnemonic.dart';
 export 'usecase_get_account.dart';
 export 'usecase_get_authentication_method.dart';
+export 'usecase_get_mnemonic.dart';
 export 'usecase_log_in.dart';
 export 'usecase_logout.dart';
 export 'usecase_refresh_account.dart';