+++ lib/usecases/account/repository_user.dart
@@ -7,8 +7,12 @@ abstract class UserRepository {
   /// allowing it to be retrieved later.
   Future<void> saveWallet(String mnemonic);
 
+  /// Returns the mnemonic that is associated to the current application user.
+  /// If no mnemonic has been saved yet, returns `null`.
+  Future<List<String>> getMnemonic();
+
   /// Returns the [Wallet] instance of the current application user.
-  /// If no [Wallet] instance has been saved yet, returns null.
+  /// If no [Wallet] instance has been saved yet, returns `null`.
   Future<Wallet> getWallet();
 
   /// Returns the [AccountData] object containing the info of the current user.